fix(ui): reload frontend after self-update from header and mobile nav
