+++ web/src/components/layout/MobileFooterNav.tsx
@@ -22,6 +22,7 @@ import { isThemePremium, themes } from "@/config/themes"
 import { useTorrentSelection } from "@/contexts/TorrentSelectionContext"
 import { useAuth } from "@/hooks/useAuth"
 import { useHasPremiumAccess } from "@/hooks/useLicense"
+import { useServerReconnect } from "@/hooks/useServerReconnect"
 import { api } from "@/lib/api"
 import { getAppVersion } from "@/lib/build-info"
 import { cn } from "@/lib/utils"
@@ -92,6 +93,7 @@ export function MobileFooterNav() {
   const { isSelectionMode } = useTorrentSelection()
   const { currentMode, currentTheme } = useThemeChange()
   const { hasPremiumAccess } = useHasPremiumAccess()
+  const { pollForReconnection, storeChangelog, ChangelogDialog } = useServerReconnect()
   const [showThemeDialog, setShowThemeDialog] = useState(false)
   const appVersion = getAppVersion()
 
@@ -108,10 +110,23 @@ export function MobileFooterNav() {
     refetchOnWindowFocus: false,
   })
 
+  // Store changelog data when update info is available
+  useEffect(() => {
+    if (updateInfo?.body) {
+      storeChangelog(updateInfo.tag_name, updateInfo.body)
+    }
+  }, [updateInfo, storeChangelog])
+
   const updateMutation = useMutation({
     mutationFn: () => api.triggerSelfUpdate(),
-    onSuccess: ({ message }) => {
-      toast.success(message)
+    onSuccess: ({ message, restart_pending }) => {
+      const notify = restart_pending ? toast.success : toast.info
+      notify(message)
+
+      if (restart_pending) {
+        // Start polling for reconnection
+        pollForReconnection()
+      }
     },
     onError: (error: unknown) => {
       const message = error instanceof Error ? error.message : "Failed to start self-update"
@@ -490,6 +505,8 @@ export function MobileFooterNav() {
           </div>
         </DialogContent>
       </Dialog>
+
+      <ChangelogDialog />
     </nav>
   )
 }
